libpressio_error_injector version 0.8.0
Major Changes

+ Now requires libpressio version 0.88.0 for pressio_thread_safety

+++ src/pressio_fault_injector.cc
@@ -72,7 +72,7 @@ class fault_injector_plugin: public libpressio_compressor_plugin {
 
   struct pressio_options 	get_configuration_impl () const override {
     pressio_options options;
-    set(options, "pressio:thread_safe", (int)pressio_thread_safety_multiple);
+    set(options, "pressio:thread_safe", pressio_thread_safety_multiple);
     set(options, "fault_injector:injection_mode_str", std::vector<std::string>{"set", "unset", "flip"});
     return options;
   };
